+++ packages/sdk/src/apiAccount.ts
@@ -680,133 +680,7 @@ export class ApiAccount<T = void> extends WithAccountHistory implements ISubmitE
     historyData?: HistoryItem,
     unsigned = false
   ): Promise<T> {
-    console.info(accountPair);
-    console.info('we are in just submitExtrinsic');
-
-    // Check if the account is a multisig account
-    const isMultisig = api.mst.getMstAccount(accountPair.address) !== undefined;
-    console.info('ismultisg', isMultisig);
-
-    if (isMultisig) {
-      // Retrieve the main account's keyring pair from storage
-      let mainAccountPair: KeyringPair | null = null;
-      if (this.accountStorage?.get('previousAccountAddress')) {
-        console.info('this.accountStorage?.get previousAccountAddress exists');
-        const previousAccountAddress = this.accountStorage?.get('previousAccountAddress');
-        mainAccountPair = keyring.getPair(previousAccountAddress);
-        console.info('mainAccountPair', mainAccountPair);
-      } else {
-        throw new Error('Main account not found in accountStorage');
-      }
-
-      if (!historyData) {
-        throw new Error('historyData is required for multisig transactions');
-      }
-
-      // Ensure we have a KeyringPair to pass in
-      if (!mainAccountPair) {
-        throw new Error('Main account keyring pair not found');
-      }
-
-      // Call submitMultisigExtrinsic with multisig account and main account key pairs
-      return await this.submitMultisigExtrinsic(
-        extrinsic,
-        accountPair, // Multisig account pair
-        mainAccountPair, // Main account pair for signing
-        historyData,
-        unsigned
-      );
-    } else {
-      // Existing logic for non-multisig accounts
-      return await this.submitApiExtrinsic(this.api, extrinsic, accountPair, this.signer, historyData, unsigned);
-    }
-  }
-
-  public async submitMultisigExtrinsic(
-    call: SubmittableExtrinsic<'promise'>,
-    multisigAccountPair: KeyringPair,
-    signerAccountPair: KeyringPair,
-    historyData: HistoryItem,
-    unsigned = false
-  ): Promise<T> {
-    console.info('we are in submitMultisigExtrinsic');
-
-    // Generate the call hash
-    const callHash = call.method.hash;
-
-    // Get multisig account and meta data
-    const multisigAccount = api.mst.getMstAccount(multisigAccountPair.address);
-    if (!multisigAccount) {
-      throw new Error(`Multisig account with address ${multisigAccountPair.address} not found.`);
-    }
-
-    // Cast meta to MultisigMeta
-    const meta = multisigAccount.meta as unknown as any;
-    if (!meta) {
-      throw new Error(`Meta data for multisig account ${multisigAccountPair.address} is missing.`);
-    }
-
-    console.info('the meta is');
-    console.info(meta);
-
-    // Get all signatories and other signatories
-    const allSignatories = meta.who ? [...meta.who] : [];
-    if (!Array.isArray(allSignatories)) {
-      throw new Error(`Signatories for multisig account ${multisigAccountPair.address} are invalid.`);
-    }
-
-    console.info('allSignatories');
-    console.info(allSignatories);
-
-    console.info('here is signerAccountPair', signerAccountPair);
-    console.info('formatted address', this.formatAddress(signerAccountPair.address));
-    const otherSignatories = allSignatories.filter(
-      (address) => address !== this.formatAddress(signerAccountPair.address)
-    );
-    otherSignatories.sort();
-
-    const threshold = meta.threshold;
-
-    console.info('the threshold', threshold);
-
-    // Check for existing approvals
-    const multisigAddress = multisigAccountPair.address;
-    const info = await this.api.query.multisig.multisigs(multisigAddress, callHash);
-    console.info('here is the info');
-    console.info(info);
-    let maybeTimepoint = null;
-
-    if (info.isSome) {
-      const { when } = info.unwrap();
-      maybeTimepoint = when;
-    }
-
-    const MAX_WEIGHT = new BN('640000000');
-    const maxWeight = this.api.registry.createType('Weight', {
-      refTime: MAX_WEIGHT,
-      proofSize: new BN(0),
-    });
-
-    const approveExtrinsic = this.api.tx.multisig.asMulti(
-      threshold,
-      otherSignatories,
-      maybeTimepoint,
-      call, // Replace `callHash` with the actual `call` (transaction payload)
-      maxWeight
-    );
-
-    // Prepare history data with 'from' address as the multisig account
-    const updatedHistoryData = { ...historyData, from: multisigAccountPair.address };
-
-    // Sign and submit the approveAsMulti extrinsic using the main account's key pair
-    return await this.submitApiExtrinsic(
-      this.api,
-      approveExtrinsic,
-      signerAccountPair, // Use main account for signing
-      this.signer,
-      updatedHistoryData,
-      unsigned
-    );
+    return await this.submitApiExtrinsic(this.api, extrinsic, accountPair, this.signer, historyData, unsigned);
   }
 
   public async signExtrinsic(
